Remove deprecated input types xmin,ymin,xmax,ymax,projection. (#77)
* Remove deprecated input types xmin,ymin,xmax,ymax,projection. 
Remove action,enctype,method attributes.
Update description of <extent> element.  

Resolves #76

Overal, description of <extent> still needs improvement, 
because since the removal of the action attribute, the meaning of 
extent has shifted slightly.  Now that the requests for content are 
driven by link templates and associated inputs, an extent is represented
the union of the extents of the templates, which is dictated by the min/
max values of the axes of the location inputs.

* Get rid of reference to obsolete action attribute.