Add geojson2mapml specification (#245)
* First pass at geojson2mapml, still unanswered questions about how to 
represent overloaded dictionary members, e.g. within featureCaptionInterface
the value provided should be a property name or a function.  How to 
represent in WebIDL, or is it even possible.  Current state is not raising
an exception, but does it mean what we want it to mean, unsure?

* Use stringifier for FeatureCaptionInterface

* Update revision history.  Add prose for back(), forward(), reload().
Add draft WebIDL for geojson2mapml() options dictionary.

* Add schema and prose for geojson2mapml and options.

* Fix return type for geometryFunction. 
Add HTMLGeometryChildElement abstract interface type.

* Add HTMLGeometryChildElement change history entry